requirements/base.txt
flask==2.1.3
gevent==21.12.0
requests==2.28.1
flasgger==0.9.5
redis==4.3.4
pydantic==1.9.0
python-dotenv==0.20.0
bcrypt==3.2.2
PyJWT==2.4.0

Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
marshmallow==3.17.0
psycopg2-binary==2.9.3
SQLAlchemy==1.4.39
flask-restx==0.5.1
Werkzeug==2.1.2
alembic==1.8.1
flask_jwt_extended==4.4.2

